+++ apps/namadillo/src/App/Common/GasFeeModal.tsx
@@ -1,4 +1,3 @@
-import { Asset } from "@chain-registry/types";
 import {
   ActionButton,
   AmountInput,
@@ -7,7 +6,7 @@ import {
 } from "@namada/components";
 import { transparentBalanceAtom } from "atoms/accounts";
 import { shieldedBalanceAtom } from "atoms/balance";
-import { chainAssetsMapAtom, nativeTokenAddressAtom } from "atoms/chain";
+import { nativeTokenAddressAtom } from "atoms/chain";
 import { GasPriceTable, GasPriceTableItem } from "atoms/fees/atoms";
 import { tokenPricesFamily } from "atoms/prices/atoms";
 import BigNumber from "bignumber.js";
@@ -16,7 +15,7 @@ import { TransactionFeeProps } from "hooks/useTransactionFee";
 import { useAtomValue } from "jotai";
 import { IoClose } from "react-icons/io5";
 import { twMerge } from "tailwind-merge";
-import { GasConfig } from "types";
+import { Asset, GasConfig, NamadaAsset } from "types";
 import { toDisplayAmount } from "utils";
 import { getDisplayGasFee } from "utils/gas";
 import { FiatCurrency } from "./FiatCurrency";
@@ -26,23 +25,23 @@ import { TokenCurrency } from "./TokenCurrency";
 const useSortByNativeToken = () => {
   const nativeToken = useAtomValue(nativeTokenAddressAtom).data;
   return (a: GasPriceTableItem, b: GasPriceTableItem) =>
-    a.token === nativeToken ? -1
-    : b.token === nativeToken ? 1
+    a.token.address === nativeToken ? -1
+    : b.token.address === nativeToken ? 1
     : 0;
 };
 
 const useBuildGasOption = ({
   gasConfig,
   gasPriceTable,
+  chainAssetsMap,
 }: {
   gasConfig: GasConfig;
   gasPriceTable: GasPriceTable | undefined;
+  chainAssetsMap: Record<string, NamadaAsset>;
 }) => {
-  const chainAssetsMap = useAtomValue(chainAssetsMapAtom);
-
   const gasDollarMap =
     useAtomValue(
-      tokenPricesFamily(gasPriceTable?.map((item) => item.token) ?? [])
+      tokenPricesFamily(gasPriceTable?.map((item) => item.token.address) ?? [])
     ).data ?? {};
 
   return (
@@ -95,10 +94,12 @@ const useBuildGasOption = ({
 export const GasFeeModal = ({
   feeProps,
   onClose,
+  chainAssetsMap,
   isShielded = false,
 }: {
   feeProps: TransactionFeeProps;
   onClose: () => void;
+  chainAssetsMap: Record<string, NamadaAsset>;
   isShielded?: boolean;
 }): JSX.Element => {
   const {
@@ -110,7 +111,11 @@ export const GasFeeModal = ({
   } = feeProps;
 
   const sortByNativeToken = useSortByNativeToken();
-  const buildGasOption = useBuildGasOption({ gasConfig, gasPriceTable });
+  const buildGasOption = useBuildGasOption({
+    gasConfig,
+    gasPriceTable,
+    chainAssetsMap,
+  });
   const nativeToken = useAtomValue(nativeTokenAddressAtom).data;
   const transparentAmount = useAtomValue(transparentBalanceAtom);
   const shieldedAmount = useAtomValue(shieldedBalanceAtom);
@@ -121,19 +126,19 @@ export const GasFeeModal = ({
       isShielded ?
         shieldedAmount.data?.map((balance) => ({
           minDenomAmount: balance.minDenomAmount,
-          tokenAddress: balance.address,
+          token: balance.address,
         }))
       : transparentAmount.data;
 
     return new BigNumber(
-      balances?.find((token) => token.tokenAddress === tokenAddres)
-        ?.minDenomAmount || "0"
+      balances?.find((token) => token.token === tokenAddres)?.minDenomAmount ||
+        "0"
     );
   };
 
   const filterAvailableTokensOnly = (item: GasPriceTableItem): boolean => {
-    if (item.token === nativeToken) return true; // we should always keep the native token
-    return findUserBalanceByTokenAddress(item.token).gt(0);
+    if (item.token.address === nativeToken) return true; // we should always keep the native token
+    return findUserBalanceByTokenAddress(item.token.address).gt(0);
   };
 
   const isLoading = isShielded && !shieldedAmount.data;
@@ -247,17 +252,17 @@ export const GasFeeModal = ({
                   totalInDollars,
                   unitValueInDollars,
                 } = buildGasOption({
-                  gasPriceInMinDenom: item.gasPrice,
-                  gasToken: item.token,
+                  gasPriceInMinDenom: item.gasPriceInMinDenom,
+                  gasToken: item.token.address,
                 });
 
                 const availableAmount = toDisplayAmount(
                   asset,
-                  findUserBalanceByTokenAddress(item.token)
+                  findUserBalanceByTokenAddress(item.token.address)
                 );
 
                 return {
-                  id: item.token,
+                  id: item.token.address,
                   value: (
                     <div
                       className={clsx(
@@ -265,7 +270,7 @@ export const GasFeeModal = ({
                         "justify-between w-full min-h-[42px] mr-5"
                       )}
                     >
-                      <TokenCard address={item.token} asset={asset} />
+                      <TokenCard address={item.token.address} asset={asset} />
                       <div>
                         <div className="text-white text-sm text-right">
                           {unitValueInDollars && (