+++ apps/namadillo/src/atoms/chain/services.ts
@@ -1,13 +1,12 @@
 import {
   DefaultApi,
   IbcToken,
+  ModelParameters,
   NativeToken,
-  Parameters,
 } from "@namada/indexer-client";
 import { getDenomFromIbcTrace } from "atoms/integrations";
 import BigNumber from "bignumber.js";
-import { findAssetByDenom } from "integrations/utils";
-import { MaspAssetRewards } from "types";
+import { MaspAssetRewards, NamadaAsset } from "types";
 import { unknownAsset } from "utils/assets";
 import { getSdkInstance } from "utils/sdk";
 
@@ -19,7 +18,7 @@ export const fetchRpcUrlFromIndexer = async (
 };
 
 // TODO: We need the response type of this call updated in the indexer client
-type TempChainParams = Parameters & {
+type TempChainParams = ModelParameters & {
   checksums: {
     current: Record<string, string>;
     fallback: Record<string, string>;
@@ -43,14 +42,17 @@ export const clearShieldedContext = async (chainId: string): Promise<void> => {
   await sdk.getMasp().clearShieldedContext(chainId);
 };
 
-export const fetchMaspRewards = async (): Promise<MaspAssetRewards[]> => {
+export const fetchMaspRewards = async (
+  assets: NamadaAsset[]
+): Promise<MaspAssetRewards[]> => {
   const sdk = await getSdkInstance();
   const rewards = await sdk.rpc.globalShieldedRewardForTokens();
   const existingRewards: MaspAssetRewards[] = rewards
     .filter((r) => r.maxRewardRate > 0)
     .map((r) => {
-      const denom = getDenomFromIbcTrace(r.name);
-      const asset = findAssetByDenom(denom) ?? unknownAsset(denom);
+      const asset =
+        assets.find((asset) => asset.address === r.address) ??
+        unknownAsset(getDenomFromIbcTrace(r.name));
       return {
         asset,
         address: r.address,