+++ apps/namadillo/src/atoms/fees/atoms.ts
@@ -1,21 +1,24 @@
-import { GasEstimate } from "@namada/indexer-client";
+import {
+  ApiV1ChainTokenGet200ResponseInner,
+  GasEstimate,
+} from "@namada/indexer-client";
 import { defaultAccountAtom } from "atoms/accounts";
 import { indexerApiAtom } from "atoms/api";
-import { chainAssetsMapAtom } from "atoms/chain";
+import { namadaRegistryChainAssetsMapAtom } from "atoms/integrations";
 import { queryDependentFn } from "atoms/utils";
 import BigNumber from "bignumber.js";
+import invariant from "invariant";
 import { atomWithQuery } from "jotai-tanstack-query";
 import { atomFamily } from "jotai/utils";
 import { isPublicKeyRevealed } from "lib/query";
 import isEqual from "lodash.isequal";
-import { Address } from "types";
 import { TxKind } from "types/txKind";
 import { isNamadaAsset, toDisplayAmount } from "utils";
 import { fetchGasEstimate, fetchTokensGasPrice } from "./services";
 
 export type GasPriceTableItem = {
-  token: Address;
-  gasPrice: BigNumber;
+  token: ApiV1ChainTokenGet200ResponseInner;
+  gasPriceInMinDenom: BigNumber;
 };
 
 export type GasPriceTable = GasPriceTableItem[];
@@ -55,18 +58,20 @@ export const gasEstimateFamily = atomFamily(
 
 export const gasPriceTableAtom = atomWithQuery<GasPriceTable>((get) => {
   const api = get(indexerApiAtom);
-  const chainAssetsMap = get(chainAssetsMapAtom);
+  const chainAssetsMap = get(namadaRegistryChainAssetsMapAtom);
 
   return {
-    queryKey: ["gas-price-table", chainAssetsMap],
+    queryKey: ["gas-price-table", chainAssetsMap.data],
     ...queryDependentFn(async () => {
+      invariant(chainAssetsMap.data, "No chain settings");
+
       const response = await fetchTokensGasPrice(api);
       return (
         response
           // filter only tokens that exists on the chain
-          .filter(({ token }) => Boolean(chainAssetsMap[token]))
+          .filter(({ token }) => Boolean(chainAssetsMap.data[token.address]))
           .map(({ token, minDenomAmount }) => {
-            const asset = chainAssetsMap[token];
+            const asset = chainAssetsMap.data[token.address];
             const baseAmount = BigNumber(minDenomAmount);
             return {
               token,
@@ -74,10 +79,11 @@ export const gasPriceTableAtom = atomWithQuery<GasPriceTable>((get) => {
                 asset && isNamadaAsset(asset) ?
                   toDisplayAmount(asset, baseAmount)
                 : baseAmount,
+              gasPriceInMinDenom: baseAmount,
             };
           })
       );
-    }, []),
+    }, [chainAssetsMap]),
   };
 });