+++ apps/namadillo/src/lib/transactions.ts
@@ -1,4 +1,3 @@
-import { Asset } from "@chain-registry/types";
 import { coin } from "@cosmjs/proto-signing";
 import {
   Attribute,
@@ -7,12 +6,13 @@ import {
 } from "@cosmjs/stargate";
 import namada from "@namada/chains/chains/namada";
 import {
-  ShieldedTransferMsgValue,
-  ShieldingTransferMsgValue,
-  TransparentTransferMsgValue,
-  UnshieldingTransferMsgValue,
-} from "@namada/types";
+  ShieldedTransferProps,
+  ShieldingTransferProps,
+  TransparentTransferProps,
+  UnshieldingTransferProps,
+} from "@namada/sdk-multicore";
 import BigNumber from "bignumber.js";
+import { Asset } from "types";
 
 import {
   allTransferStages,
@@ -195,10 +195,10 @@ export const createTransferDataFromNamada = (
   rpcUrl: string,
   isShieldedTx: boolean,
   txResponse?:
-    | TransactionPair<TransparentTransferMsgValue>
-    | TransactionPair<ShieldedTransferMsgValue>
-    | TransactionPair<ShieldingTransferMsgValue>
-    | TransactionPair<UnshieldingTransferMsgValue>,
+    | TransactionPair<TransparentTransferProps>
+    | TransactionPair<ShieldedTransferProps>
+    | TransactionPair<ShieldingTransferProps>
+    | TransactionPair<UnshieldingTransferProps>,
   memo?: string
 ): TransferTransactionData[] => {
   if (!txResponse?.encodedTxData?.txs?.length) {