+++ apps/namadillo/src/types.ts
@@ -1,11 +1,12 @@
-import { Asset, AssetList, Chain, IBCInfo } from "@chain-registry/types";
-import { ValidatorStatus } from "@namada/indexer-client";
 import {
-  Account,
-  ChainKey,
-  ClaimRewardsMsgValue,
-  ExtensionKey,
-} from "@namada/types";
+  AssetList,
+  Chain,
+  Asset as ChainRegistryAsset,
+  IBCInfo,
+} from "@chain-registry/types";
+import { ValidatorStatus } from "@namada/indexer-client";
+import { ClaimRewardsMsgValue } from "@namada/sdk-multicore";
+import { Account, ChainKey, ExtensionKey } from "@namada/types";
 import { MutationStatus } from "@tanstack/query-core";
 import BigNumber from "bignumber.js";
 import { TxKind } from "types/txKind";
@@ -213,13 +214,28 @@ export type ChainRegistryEntry = {
   ibc?: IBCInfo[];
 };
 
-export type AddressWithAsset = {
-  originalAddress: Address;
-  asset: Asset;
+export type NamadaChainRegistryEntry = ChainRegistryEntry & {
+  assets: AssetList & { assets: NamadaAsset[] };
 };
 
-export type AddressWithAssetAndAmount = AddressWithAsset & {
+export type Asset = ChainRegistryAsset;
+
+// Namada assets should always have address field defined
+export type NamadaAsset = Asset & { address: Address };
+
+export type AssetWithAmount = {
+  asset: Asset;
   amount: BigNumber;
+};
+
+export type NamadaAssetWithAmount = {
+  asset: NamadaAsset;
+  amount: BigNumber;
+};
+
+export type AssetWithMinDenomAmount = {
+  asset: Asset;
+  minDenomAmount: BigNumber;
 };
 
 export type Coin = {
@@ -227,10 +243,17 @@ export type Coin = {
   minDenomAmount: string;
 };
 
-export type AddressWithAssetAndAmountMap = Record<
-  Address,
-  AddressWithAssetAndAmount
->;
+export type TokenBalance = {
+  address: Address;
+  asset: Asset;
+  amount: BigNumber;
+  dollar?: BigNumber;
+};
+
+export type IbcChannels = {
+  namadaChannel: string;
+  ibcChannel: string;
+};
 
 export enum TransferStep {
   Sign = "sign",