+++ apps/namadillo/src/utils/index.ts
@@ -1,16 +1,16 @@
-import { Asset, DenomUnit } from "@chain-registry/types";
-import namadaAssets from "@namada/chain-registry/namada/assetlist.json";
+import { DenomUnit } from "@chain-registry/types";
+import { TxProps } from "@namada/sdk-multicore";
 import {
   BroadcastTxError,
   ProposalStatus,
-  ProposalTypeString,
+  ProposalType,
   ResultCode,
-  TxMsgValue,
 } from "@namada/types";
-import { localnetConfigAtom } from "atoms/integrations/atoms";
+import { getNamadaChainAssetsMap } from "atoms/integrations/functions";
 import BigNumber from "bignumber.js";
-import { getDefaultStore } from "jotai";
+import invariant from "invariant";
 import { useEffect, useRef } from "react";
+import { Asset } from "types";
 
 export const proposalStatusToString = (status: ProposalStatus): string => {
   const statusText: Record<ProposalStatus, string> = {
@@ -23,10 +23,8 @@ export const proposalStatusToString = (status: ProposalStatus): string => {
   return statusText[status];
 };
 
-export const proposalTypeStringToString = (
-  type: ProposalTypeString
-): string => {
-  const typeText: Record<ProposalTypeString, string> = {
+export const proposalTypeStringToString = (type: ProposalType): string => {
+  const typeText: Record<ProposalType, string> = {
     default: "Default",
     default_with_wasm: "Default with Wasm",
     pgf_steward: "PGF Steward",
@@ -78,23 +76,12 @@ const findDisplayUnit = (asset: Asset): DenomUnit | undefined => {
 };
 
 export const namadaAsset = (): Asset => {
-  const store = getDefaultStore();
-  const config = store.get(localnetConfigAtom);
+  // This works for both housefire and mainnet because the native asset is the same
+  const namadaAssets = Object.values(getNamadaChainAssetsMap(false));
+  const nativeAsset = namadaAssets.find((asset) => asset.base === "unam");
+  invariant(nativeAsset, "Namada native asset not found");
 
-  const configTokenAddress = config.data?.tokenAddress;
-
-  // TODO we should get this dynamically from the Github like how we do for chains
-  const assets = namadaAssets.assets as Asset[];
-  const registryAsset = assets[0];
-  const asset =
-    configTokenAddress ?
-      {
-        ...registryAsset,
-        address: configTokenAddress,
-      }
-    : registryAsset;
-
-  return asset;
+  return nativeAsset;
 };
 
 export const isNamadaAsset = (asset?: Asset): boolean =>
@@ -131,7 +118,7 @@ export const toGasMsg = (gasLimit: BigNumber): string => {
  * Returns formatted error message based on tx props and error code
  */
 export const toErrorDetail = (
-  tx: TxMsgValue[],
+  tx: TxProps[],
   error: BroadcastTxError
 ): string => {
   try {
@@ -155,7 +142,7 @@ export const toErrorDetail = (
   }
 };
 
-export const textToErrorDetail = (text: string, tx: TxMsgValue): string => {
+export const textToErrorDetail = (text: string, tx: TxProps): string => {
   const { args } = tx;
 
   if (text.includes("Gas error:")) {